refactor: simplify build and packaging of breadcrumb component
