+++ packages/components/breadcrumb/css/index.js
@@ -1,4 +1,2 @@
-export { BsBreadcrumb } from "./bs-breadcrumb.js";
 export { BsBreadcrumbCss } from "./bs-breadcrumb.css.js";
-export { BsBreadcrumbItem } from "./bs-breadcrumb-item.js";
 export { BsBreadcrumbItemCss } from "./bs-breadcrumb-item.css.js";
