--- packages/components/breadcrumb/custom-elements.json
@@ -1,132 +1,0 @@
-{
-  "version": 2,
-  "tags": [
-    {
-      "name": "bs-breadcrumb-item",
-      "description": "Breadcrumb item",
-      "jsDoc": "/**\n * Breadcrumb item\n *\n * @element bs-breadcrumb-item\n *\n * @attribute title - The breadcrub title\n * @attribute active - Indicates if this is the last breadcrumb item\n * @attribute href - The hyper link to point to (not applicable if last breadcrumb item)\n * @attribute target - The context target (not applicable if last breadcrumb item)\n *\n * @property {String} title - Indicates whether or not the alert can be dismissed. default: false\n * @property {String} active - Indicates if this is the last breadcrumb item. default: false\n * @property {String} href - The hyper link to point to (not applicable if last breadcrumb item). default: ''\n * @property {String} target - The context target (not applicable if last breadcrumb item). default: _self\n *\n * @cssproperty --breadcrumb-item-padding-left Left padding. default: 0.5rem\n * @cssproperty --breadcrumb-item-active-color Text color when active attribute is used. default: #6c757d\n * @cssproperty --breadcrumb-item-divider-content Divider content. default: \"/\"\n * @cssproperty --breadcrumb-item-divider-color Divider color. default: #6c757d\n * @cssproperty --breadcrumb-item-divider-display Divider display. default: inline-block\n * @cssproperty --breadcrumb-item-divider-padding-right Divider right padding. default: 0.3rem\n * @cssproperty --breadcrumb-item-divider-text-decoration Divider text decoration. default: none\n * @cssproperty --breadcrumb-item-link-color Link text color. default: #007bff\n * @cssproperty --breadcrumb-item-link-text-decoration Link text decoration. default: none\n * @cssproperty --breadcrumb-item-link-bg-color Link background color. default: transparent\n * @cssproperty --breadcrumb-link-item-link-hover-color Link hover text color. default: #0056b3\n * @cssproperty --breadcrumb-item-link-hover-text-decoration Link hover text decoration. default: underline\n */",
-      "attributes": [
-        {
-          "name": "title",
-          "type": "string"
-        },
-        {
-          "name": "href",
-          "type": "string"
-        },
-        {
-          "name": "target",
-          "type": "string"
-        },
-        {
-          "name": "active",
-          "type": "boolean"
-        }
-      ],
-      "properties": [
-        {
-          "name": "title",
-          "type": "string"
-        },
-        {
-          "name": "href",
-          "type": "string"
-        },
-        {
-          "name": "target",
-          "type": "string"
-        },
-        {
-          "name": "active",
-          "type": "boolean"
-        }
-      ],
-      "cssProperties": [
-        {
-          "name": "--breadcrumb-item-link-hover-text-decoration Link hover text decoration. default: underline"
-        },
-        {
-          "name": "--breadcrumb-link-item-link-hover-color Link hover text color. default: #0056b3"
-        },
-        {
-          "name": "--breadcrumb-item-link-bg-color Link background color. default: transparent"
-        },
-        {
-          "name": "--breadcrumb-item-link-text-decoration Link text decoration. default: none"
-        },
-        {
-          "name": "--breadcrumb-item-link-color Link text color. default: #007bff"
-        },
-        {
-          "name": "--breadcrumb-item-divider-text-decoration Divider text decoration. default: none"
-        },
-        {
-          "name": "--breadcrumb-item-divider-padding-right Divider right padding. default: 0.3rem"
-        },
-        {
-          "name": "--breadcrumb-item-divider-display Divider display. default: inline-block"
-        },
-        {
-          "name": "--breadcrumb-item-divider-color Divider color. default: #6c757d"
-        },
-        {
-          "name": "--breadcrumb-item-divider-content Divider content. default: \"/\""
-        },
-        {
-          "name": "--breadcrumb-item-active-color Text color when active attribute is used. default: #6c757d"
-        },
-        {
-          "name": "--breadcrumb-item-padding-left Left padding. default: 0.5rem"
-        }
-      ]
-    },
-    {
-      "name": "bs-breadcrumb",
-      "description": "Breadcrumb Element",
-      "jsDoc": "/**\n * Breadcrumb Element\n *\n * @element bs-breadcrumb\n *\n * @slot Slot for breadcrub items content\n *\n * @cssproperty --breadcrumb-display - Display property. default: flex\n * @cssproperty --breadcrumb-flex-wrap - Flex-wrap. default: wrap\n * @cssproperty --breadcrumb-padding-top - Padding top. default: 0.75rem\n * @cssproperty --breadcrumb-padding-bottom - Padding bottom. default: 0.75rem\n * @cssproperty --breadcrumb-padding-right - Padding right. default: 1rem\n * @cssproperty --breadcrumb-padding-left - Padding left. default: 1rem\n * @cssproperty --breadcrumb-margin-bottom - Margin bottom. default: 1rem\n * @cssproperty --breadcrumb-bg-color - Background color. default: #e9ecef\n * @cssproperty --breadcrumb-border-radius - Border radius. default: 0.25rem\n */",
-      "slots": [
-        {
-          "name": "Slot for breadcrub items content"
-        }
-      ],
-      "cssProperties": [
-        {
-          "name": "--breadcrumb-border-radius",
-          "description": "Border radius. default: 0.25rem"
-        },
-        {
-          "name": "--breadcrumb-bg-color",
-          "description": "Background color. default: #e9ecef"
-        },
-        {
-          "name": "--breadcrumb-margin-bottom",
-          "description": "Margin bottom. default: 1rem"
-        },
-        {
-          "name": "--breadcrumb-padding-left",
-          "description": "Padding left. default: 1rem"
-        },
-        {
-          "name": "--breadcrumb-padding-right",
-          "description": "Padding right. default: 1rem"
-        },
-        {
-          "name": "--breadcrumb-padding-bottom",
-          "description": "Padding bottom. default: 0.75rem"
-        },
-        {
-          "name": "--breadcrumb-padding-top",
-          "description": "Padding top. default: 0.75rem"
-        },
-        {
-          "name": "--breadcrumb-flex-wrap",
-          "description": "Flex-wrap. default: wrap"
-        },
-        {
-          "name": "--breadcrumb-display",
-          "description": "Display property. default: flex"
-        }
-      ]
-    }
-  ]
-}
