+++ packages/components/breadcrumb/index.js
@@ -1,0 +1,3 @@
+export * from './css'
+export { BsBreadcrumb } from "./bs-breadcrumb.js";
+export { BsBreadcrumbItem } from "./bs-breadcrumb-item.js";
