+++ packages/components/breadcrumb/package.json
@@ -2,8 +2,8 @@
   "name": "@lit-element-bootstrap/breadcrumb",
   "version": "1.0.1",
   "description": "Bootstrap v4.3.1 breadcrumb component in LitElement",
-  "main": "src/index.js",
-  "module": "src/index.js",
+  "main": "index.js",
+  "module": "index.js",
   "type": "module",
   "author": "Nikolas Christou",
   "homepage": "https://github.com/nik-christou/lit-element-bootstrap",
@@ -12,15 +12,25 @@
     "url": "https://github.com/nik-christou/lit-element-bootstrap"
   },
   "license": "MIT",
+  "files": [
+    "index.js",
+    "index.d.ts",
+    "global.d.ts",
+    "bs-*.js",
+    "bs-*.d.ts",
+    "css/*.js",
+    "css/*.d.ts",
+    "custom-elements.json"
+  ],
   "scripts": {
-    "pika:build": "pika build",
-    "tsc:build": "rimraf ./types && tsc -p tsconfig.json",
+    "prepack": "yarn build",
+    "build": "npm run tsc:build; npm run wca-analyze:json",
+    "tsc:build": "tsc -p tsconfig.json",
     "serve:demo": "es-dev-server --node-resolve --app-index demo.html --preserve-symlinks",
-    "lit-analyze": "lit-analyzer src --failFast",
-    "wca-analyze": "wca analyze \"./src/**/*.js\"",
-    "wca-analyze:json": "wca analyze \"./src/**/*.js\" --outFile custom-elements.json",
-    "lint:prettier": "prettier \"./src/**/*.js\" --tab-width 4 --list-different",
-    "format:prettier": "prettier \"./src/**/*.js\" --tab-width 4 --write",
+    "lit-analyze": "lit-analyzer bs-*.js css/*.js --failFast",
+    "wca-analyze:json": "wca analyze . --outFile custom-elements.json",
+    "lint:prettier": "prettier \"./**/*.js\" --tab-width 4 --list-different",
+    "format:prettier": "prettier \"./**/*.js\" --tab-width 4 --write",
     "test": "karma start",
     "test:coverage": "karma start --coverage",
     "test:watch": "karma start --auto-watch=true --single-run=false",
@@ -36,10 +46,6 @@
     "@open-wc/karma-esm": "^2.9.8",
     "@open-wc/prettier-config": "^0.1.12",
     "@open-wc/testing": "^2.3.9",
-    "@pika/pack": "^0.5.0",
-    "@pika/plugin-build-web": "^0.7.1",
-    "@pika/plugin-bundle-web": "^0.7.1",
-    "@pika/plugin-standard-pkg": "^0.7.1",
     "axe-core": "^3.4.0",
     "chai": "^4.2.0",
     "es-dev-server": "^1.23.1",
@@ -51,25 +57,8 @@
     "karma-mocha": "^1.3.0",
     "karma-mocha-reporter": "^2.2.5",
     "lit-analyzer": "^1.1.9",
-    "rimraf": "^3.0.0",
     "sinon": "^7.5.0",
     "typescript": "^3.7.2",
     "web-component-analyzer": "^0.1.20"
-  },
-  "@pika/pack": {
-    "pipeline": [
-      [
-        "@pika/plugin-standard-pkg"
-      ],
-      [
-        "@pika/plugin-build-web"
-      ],
-      [
-        "@pika/plugin-bundle-web",
-        {
-            "entrypoint": "unpkg"
-        }
-      ]
-    ]
   }
 }
