--- packages/components/breadcrumb/types/bs-breadcrumb-item.css.d.ts
@@ -1,1 +1,0 @@
-export const BsBreadcrumbItemCss: import("lit-element").CSSResult;
