--- packages/components/breadcrumb/types/bs-breadcrumb-reboot.css.d.ts
@@ -1,1 +1,0 @@
-export const BsBreadcrumbRebootCss: import("lit-element").CSSResult;
