--- packages/components/breadcrumb/types/bs-breadcrumb.css.d.ts
@@ -1,1 +1,0 @@
-export const BsBreadcrumbCss: import("lit-element").CSSResult;
